Routen hinzugefügt etc

+++ app/src/main/java/com/myapp/model/room/entities/Challenges.kt
@@ -5,7 +5,7 @@ import androidx.room.PrimaryKey
 
 @Entity(tableName = "challenges")
 data class Challenges(
-    @PrimaryKey(autoGenerate = true) val id: Long? = null,
+    @PrimaryKey(autoGenerate = true) val id: Int? = null,
 
     // Titel der Challenge
     val title: String,
